+++ superset/commands/annotation_layer/delete.py
@@ -39,7 +39,8 @@ class DeleteAnnotationLayerCommand(BaseCommand):
     @transaction(on_error=partial(on_error, reraise=AnnotationLayerDeleteFailedError))
     def run(self) -> None:
         self.validate()
-        assert self._models
+        if not self._models:
+            raise RuntimeError("Models were not set by validate")
         AnnotationLayerDAO.delete(self._models)
 
     def validate(self) -> None:
