+++ superset/commands/css/delete.py
@@ -38,7 +38,8 @@ class DeleteCssTemplateCommand(BaseCommand):
     @transaction(on_error=partial(on_error, reraise=CssTemplateDeleteFailedError))
     def run(self) -> None:
         self.validate()
-        assert self._models
+        if not self._models:
+            raise RuntimeError("Models were not set by validate")
         CssTemplateDAO.delete(self._models)
 
     def validate(self) -> None:
