+++ superset/commands/query/delete.py
@@ -38,7 +38,8 @@ class DeleteSavedQueryCommand(BaseCommand):
     @transaction(on_error=partial(on_error, reraise=SavedQueryDeleteFailedError))
     def run(self) -> None:
         self.validate()
-        assert self._models
+        if not self._models:
+            raise RuntimeError("Models were not set by validate")
         SavedQueryDAO.delete(self._models)
 
     def validate(self) -> None:
