+++ superset/commands/report/delete.py
@@ -41,7 +41,8 @@ class DeleteReportScheduleCommand(BaseCommand):
     @transaction(on_error=partial(on_error, reraise=ReportScheduleDeleteFailedError))
     def run(self) -> None:
         self.validate()
-        assert self._models
+        if not self._models:
+            raise RuntimeError("Models were not set by validate")
         ReportScheduleDAO.delete(self._models)
 
     def validate(self) -> None:
